% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.glmpca_fit}
\alias{fitted.glmpca_fit}
\title{Get Fitted Values for GLM-PCA Model Fit}
\usage{
\method{fitted}{glmpca_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \dQuote{glmpca_fit},
typically the result of calling \code{\link{fit_glmpca}}.}
}
\value{
An n x p matrix of fitted means. Calculated as
  \deqn{exp(L'F)} using the \code{fit} object.
}
\description{
\code{fitted} method for the
  \dQuote{glmpca_fit} class.
}
