% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_glmpca_binom}
\alias{init_glmpca_binom}
\title{Initialize Binomial GLM-PCA Fit}
\usage{
init_glmpca_binom(Y, K, LL, FF, fixed_loadings = NULL, fixed_factors = NULL)
}
\arguments{
\item{Y}{The n x p matrix of counts; all entries of Y should be
non-negative.}

\item{K}{An integer 1 or greater giving the matrix rank. This
argument should only be specified if the initial fit (\code{LL, FF})
is not provided.}

\item{LL}{An optional argument giving the initial estimate of the
loadings matrix. It should be an K x n matrix, where n is the
number of rows in the counts matrix \code{Y}, and K >= 1 is the rank
of the matrix factorization. When \code{LL} and \code{FF} are not
provided, input argument \code{K} should be specified instead, and
\code{LL} and \code{FF} are initialized as a constant matrix.}

\item{FF}{An optional argument giving is the initial estimate of the
factors matrix. It should be a K x p matrix, where p is the number
of columns in the counts matrix \code{Y}, and K >= 1 is the rank of
the matrix factorization. When \code{LL} and \code{FF} are not
provided, input argument \code{K} should be specified instead, and
\code{LL} and \code{FF} are initialized randomly}

\item{fixed_loadings}{Vector of integers indicating, which, if any, loadings
should be fixed at their initial values. This argument will be ignored if
\code{LL} is not provided.}

\item{fixed_factors}{Vector of integers indicating which, if any, factors
should be fixed at their initial values. This argument will be ignored
if \code{FF} is not provided.}

\item{N}{The n x p matrix of trials; all entries of N should be
at least 1.}
}
\value{
An object capturing the initial state of the model fit. See
  \code{\link{fit_glmpca}} for details.
}
\description{
Initialize a binomial GLM-PCA model fit.
}
\seealso{
\code{\link{fit_glmpca_binom}}
}
