% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_glmpca_pois_greedy}
\alias{init_glmpca_pois_greedy}
\title{Initialize GLM-PCA Poisson Fit With Greedy Fitting}
\usage{
init_glmpca_pois_greedy(Y, K, max_greedy_iter = 10)
}
\arguments{
\item{Y}{n x p matrix of counts; all entries should be
non-negative.}

\item{K}{An integer 1 or greater giving the matrix rank. This
argument should only be specified if the initial fit (\code{LL, FF})
is not provided.}

\item{max_greedy_iter}{maximum number of updates to each greedily
added loading / factor pair}
}
\value{
An object capturing the initial state of the model fit. See
  \code{\link{fit_glmpca_pois}} for details.
}
\description{
Initialize a GLM-PCA Poisson model fit and greedily
update each factor and loading
}
\seealso{
\code{\link{fit_glmpca_pois}}
}
