% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plash.R
\name{plash_omni}
\alias{plash_omni}
\title{Title}
\usage{
plash_omni(
  Y,
  K,
  offset = FALSE,
  intercept = FALSE,
  constant_offset = TRUE,
  parallel = TRUE,
  tol = 1e-08,
  update_L = T,
  update_F = T,
  update_c = T,
  update_L_c_simul = FALSE,
  init_LL = NULL,
  init_FF = NULL,
  init_cc = NULL
)
}
\arguments{
\item{update_c}{}
}
\value{

}
\description{
Title
}
