% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plash.R
\name{plash_parallel}
\alias{plash_parallel}
\title{Title}
\usage{
plash_parallel(
  Y,
  K,
  cluster,
  tol = 1e-08,
  update_L = T,
  update_F = F,
  update_c = F,
  true_LL = NULL,
  true_FF = NULL,
  true_cc = NULL
)
}
\arguments{
\item{update_c}{}
}
\value{

}
\description{
Title
}
