% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plash.R
\name{plash_parallel_glmpca}
\alias{plash_parallel_glmpca}
\title{Title}
\usage{
plash_parallel_glmpca(
  Y,
  K,
  cluster,
  offset = FALSE,
  offset_vec = NULL,
  intercept = FALSE,
  tol = 1e-08,
  update_L = T,
  update_F = T,
  update_c = F,
  true_LL = NULL,
  true_FF = NULL,
  true_cc = NULL
)
}
\arguments{
\item{update_c}{}
}
\value{

}
\description{
Title
}
