% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{set_fastglmpca_threads}
\alias{set_fastglmpca_threads}
\title{Set up Multithreading for fastglmpca}
\usage{
set_fastglmpca_threads(n)
}
\arguments{
\item{n}{The requested number of threads.}
}
\value{
The number of threads to be used.
}
\description{
Initialize RcppParallel multithreading using a
  pre-specified number of threads, or using the default number of
  threads when \code{n} is not specified or is NA.
}
