% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.glmpca_pois_fit}
\alias{summary.glmpca_pois_fit}
\alias{print.summary.glmpca_pois_fit}
\title{Summarize GLM-PCA Model Fit}
\usage{
\method{summary}{glmpca_pois_fit}(object, ...)

\method{print}{summary.glmpca_pois_fit}(x, ...)
}
\arguments{
\item{object}{An object of class \dQuote{glmpca_fit},
typically the result of calling \code{\link{fit_glmpca_pois}}.}

\item{\dots}{Additional arguments passed to the generic
\code{summary} or \code{print.summary} method.}

\item{x}{An object of class \dQuote{summary.glmpca_fit},
usually the result of a call to \code{summary.glmpca_fit}.}
}
\value{
A vector of statistics summarizing the model fit.
}
\description{
\code{summary} method for the
  \dQuote{glmpcan_fit} class.
}
