% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/init.R
\name{fit_glmpca_pois}
\alias{fit_glmpca_pois}
\alias{fit_glmpca_pois_control_default}
\alias{init_glmpca_pois}
\title{Fit Poisson GLM-PCA Model to Count Data}
\usage{
fit_glmpca_pois(
  Y,
  K,
  fit0 = init_glmpca_pois(Y, K),
  tol = 0.0001,
  min_iter = 1,
  max_iter = 100,
  verbose = TRUE,
  control = list()
)

fit_glmpca_pois_control_default()

init_glmpca_pois(
  Y,
  K,
  U,
  V,
  X = numeric(0),
  Z = numeric(0),
  B = numeric(0),
  W = numeric(0),
  fixed_b_cols = numeric(0),
  fixed_w_cols = numeric(0),
  col_size_factor = TRUE,
  row_intercept = TRUE
)
}
\arguments{
\item{Y}{The n x m matrix of counts; all entries of \code{Y} should
be non-negative. It can be a sparse matrix (class
\code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{K}{Integer 1 or greater specifying the rank of the matrix
factorization. This should only be provided if the initial fit
(\code{fit0}) is not.}

\item{fit0}{Initial model fit. It should be an object of class
\dQuote{glmpca_fit_pois}, such as an output from
\code{init_glmpca_pois} or a previous call to
\code{fit_glmpca_pois}.}

\item{tol}{The optimization steps when the change in the
log-likelihood between two successive iterations is less than this
amount.}

\item{min_iter}{Minimum nummber of updates to be performed.}

\item{max_iter}{Maximum number of updates to be performed.}

\item{verbose}{If \code{verbose = TRUE}, information about the
algorithm's progress is printed after each update.}

\item{control}{List of control parameters to modify behavior of
the optimization algorithm; see \dQuote{Details}.}

\item{U}{An optional argument giving the initial estimate of the
loadings matrix. It should be an n x K matrix, where n is the
number of rows in the counts matrix \code{Y}, and K > 0 is the rank
of the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{V}{An optional argument giving is the initial estimate of the
factors matrix. It should be a m x K matrix, where m is the number
of columns in the counts matrix \code{Y}, and K > 0 is the rank of
the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{X}{Optional argument giving row covariates of the count
matrix \code{Y}. It should be an n x nx matrix.}

\item{Z}{Optional argument giving column covariates of the count
matrix \code{Y}. It should be an m x nz matrix.}

\item{B}{Optional argument giving the initial estimates for the
coefficients of the row covariates. It should be an m x nx matrix.
This argument is ignored if X is not provided.}

\item{W}{Optional argument giving the initial estimates for the
coefficients of the column covariates.  It should be an n x nz
matrix. This argument is ignored if Z is not provided.}

\item{fixed_b_cols}{Optional numeric vector specifying which
columns of \code{B} (if any) should be fixed during
optimization. This argument is ignored if X is not provided.}

\item{fixed_w_cols}{Optional numeric vector specifying which
columns of \code{W} (if any) should be fixed during
optimization. This argument is ignored if Z is not provided.}

\item{col_size_factor}{If \code{col_size_factor = TRUE}, add a
fixed factor accounting for average differences in Poisson rates
across columns of \code{Y}. Setting \code{col_size_factor = TRUE}
and \code{row_intercept = TRUE} is intended to replicate what
\code{glmpca} does.}

\item{row_intercept}{If \code{row_intercept = TRUE}, add a fixed
factor accounting for average differences in Poisson rates across
rows of \code{Y}. Setting \code{col_size_factor = TRUE}
and \code{row_intercept = TRUE} is intended to replicate what
\code{glmpca} does.}
}
\value{
An object capturing the final state of the model fit. It
  contains the final estimates of \eqn{U}, \eqn{D} and \eqn{V}, and
  the other parameters (\eqn{X}, \eqn{B}, \eqn{Z}, \eqn{W}), as well
  as the log-likelihood achieved (\code{loglik}), and a data frame
  \code{progress} storing information about the algorithm's progress
  after each update.
}
\description{
Fit a Poisson GLM-PCA model by maximum-likelihood.
}
\details{
In generalized principal component analysis (GLM-PCA)
based on a Poisson likelihood, the counts
\eqn{y_{ij}} stored in an n x m matrix \eqn{Y} are modeled as
\deqn{y_{ij} \sim Pois(\lambda_{ij}),} in which 
the logarithm of each rate parameter \eqn{\lambda_{ij}}
is defined as a linear combination of rank-K matrices to be
estimated from the data:
\deqn{\log \lambda_{ij} = (UDV')_{ij},}
where \eqn{U} and \eqn{V} are orthogonal matrices of dimension
\eqn{n x K} and \eqn{m x K}, respectively, and \eqn{D} is a
diagonal \eqn{K x K} matrix in which the entries along its diagonal
are positive and decreasing.  \eqn{K} is a tuning parameter
specifying the rank of the matrix factorization. This is the same
as the low-rank matrix decomposition underlying PCA (that is, the
singular value decomposition), but because we are not using a
linear (Gaussian) model, this called \dQuote{generalized PCA} or
\dQuote{GLM PCA}.

To allow for additional components that may be fixed,
\code{fit_glmpca_pois} can also fit the more general model
\deqn{\log \lambda_{ij} = (UDV' + XB' + WZ')_{ij},}
in which \eqn{X}, \eqn{Z} are fixed matrices of dimension \eqn{n x
nx} and \eqn{m x nz}, respectively, and \eqn{B}, \eqn{W} are
matrices of dimension \eqn{m x nx} and \eqn{n x nz} to be estimated
from the data.

\code{fit_glmpca_pois} computes maximum-likelihood estimates (MLEs)
of \eqn{U}, \eqn{V}, \eqn{D}, \eqn{B} and \eqn{W} satistifying the
orthogonality constraints for \eqn{U} and \eqn{V} and the
additional constraints on \eqn{D} that the entries are positive and
decreasing. This is accomplished by iteratively fitting a series of
Poisson GLMs, and each of these individual Poissons GLMs is fitted
using a fast cyclic co-ordinate descent (CCD) algorithm.

The \code{control} argument is a list in which any of the following
named components will override the default optimization algorithm
settings (as they are defined by \code{fit_glmpca_pois_control_default}):

\describe{

\item{\code{num_ccd_iter}}{Number of CD updates to be made to
  parameters at each iteration of the algorithm.}

\item{\code{line_search}}{If \code{line_search = TRUE},
  backtracking line search performed at each iteration of CCD to
  guarantee improvement in the objective (the log-likelihood).}

\item{\code{alpha}}{alpha parameter for backtracking line search.
  (Should be a number between 0 and 0.5, typically a number near
  zero.)}

\item{\code{beta}}{beta parameter for backtracking line search
  controlling the rate at which the step size is decreased.
  (Should be a number between 0 and 0.5.)}
  
\item{\code{calc_deriv}}{If \code{calc_deriv = TRUE}, the maximum
  gradient of \eqn{U} and \eqn{V} is calculated and stored after each
  update. This may be useful for assessing convergence of the
  optimization, though increases overhead.}
  
\item{\code{calc_max_diff}}{If \code{calc_max_diff = TRUE}, the
  largest change in \eqn{U} and \eqn{V} after each update is
  calculated and stored. This may be useful for monitoring progress
  of the optimization algorithm.}
}
}
\examples{
set.seed(1)

n <- 1000
p <- 500
K <- 1

data <- generate_glmpca_data_pois(n, p, K)
# initialize fit
fit0 <- init_glmpca_pois(
  Y = data$Y, 
  K = K,
  fit_col_size_factor = TRUE, 
  fit_row_intercept = TRUE
)

fit <- fit_glmpca_pois(Y = data$Y, fit0 = fit0)

}
\references{
Townes, F. W., Hicks, S. C., Aryee, M. J. and Irizarry,
  R. A. (2019). Feature selection and dimension reduction for
  single-cell RNA-Seq based on a multinomial model. \emph{Genome Biology}
  \bold{20}, 295. \url{https://doi.org/10.1186/s13059-019-1861-6}

  Collins, M., Dasgupta, S. and Schapire, R. E. (2002). A
  generalization of principal components analysis to the exponential
  family. In \emph{Advances in Neural Information Processing Systems} 14.
}
\seealso{
\code{\link{fit_glmpca_pois}}
}
