% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/init.R
\name{fit_glmpca_pois}
\alias{fit_glmpca_pois}
\alias{init_glmpca_pois}
\title{Fit Poisson GLM-PCA Model to Count Data}
\usage{
fit_glmpca_pois(
  Y,
  K,
  fit0,
  tol = 0.0001,
  min_iter = 1,
  max_iter = 100,
  verbose = TRUE,
  control = list()
)

init_glmpca_pois(
  Y,
  K,
  U,
  V,
  X = numeric(0),
  Z = numeric(0),
  B = numeric(0),
  W = numeric(0),
  fixed_b_cols = numeric(0),
  fixed_w_cols = numeric(0),
  col_size_factor = TRUE,
  row_intercept = TRUE,
  row_size_factor = FALSE,
  col_intercept = FALSE
)
}
\arguments{
\item{Y}{The n x m matrix of counts; all entries of \code{Y} should
be non-negative. It can be a sparse matrix (class
\code{"dgCMatrix"}) or dense matrix (class \code{"matrix"}).}

\item{K}{Integer 1 or greater giving the rank of the matrix
factorization. This argument should only be specified if 
initial estimates \code{U}, \code{V} are not provided.}

\item{fit0}{The initial model fit. It should be an object of class
\dQuote{glmpca_fit}, such as an output from \code{init_glmpca_pois}, 
or from a previous call to \code{fit_glmpca}.}

\item{tol}{Positive scalar determining relative tolerance for assessing convergence.
Convergence is determined by comparing the log-likelihood at the previous
iteration to the current iteration.}

\item{min_iter}{Minimum number of updates to 
\eqn{U}, \eqn{V}, \eqn{W}, and \eqn{B} to be run.}

\item{max_iter}{Maximum number of updates to 
\eqn{U}, \eqn{V}, \eqn{W}, and \eqn{B} to be run.}

\item{verbose}{Boolean indicating if likelihood should be printed at each step.}

\item{control}{List of control parameters to modify behavior of \code{algorithm}.}

\item{U}{An optional argument giving the initial estimate of the
loadings matrix. It should be an n x K matrix, where n is the
number of rows in the counts matrix \code{Y}, and K > 0 is the rank
of the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{V}{An optional argument giving is the initial estimate of the
factors matrix. It should be a m x K matrix, where m is the number
of columns in the counts matrix \code{Y}, and K > 0 is the rank of
the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{X}{Optional argument giving row covariates of the count
matrix \code{Y}. It should be an n x nx matrix, where n is the
number of rows in \code{Y} and nx is the number of row
covariates.}

\item{Z}{An optional argument giving column specific covariates of the 
count matrix \code{Y}. It should be p x n_z matrix, where p
is the number of columns of the count matrix \code{Y}, and
n_z is the number of column covariates.}

\item{B}{An optional argument giving the initial estimates
for the coefficients of the row specific covariates of the 
count matrix \code{Y}. It should be p x n_x matrix, where p
is the number of columns of the count matrix \code{Y}, and
n_x is the number of row covariates. This argument is ignored
if X is not provided.}

\item{W}{An optional argument giving the initial estimates
for the coefficients of the column specific covariates of the 
count matrix \code{Y}. It should be n x n_z matrix, where n
is the number of rows of the count matrix \code{Y}, and
n_z is the number of column covariates.}

\item{fixed_b_cols}{An optional argument giving which, if any, columns
of \code{B} should be fixed during optimization.}

\item{fixed_w_cols}{An optional argument giving which, if any, columns
of \code{W} should be fixed during optimization.}

\item{col_size_factor}{Boolean indicating if a size factor should be
used to normalize the likelihood across columns. This may be
useful when \code{Y} is a matrix from a scRNA experiment where rows 
represent genes and columns represent cells.}

\item{row_intercept}{Boolean indicating if intercept term should be
fit for each row of \code{Y}. This may be useful when \code{Y} is a
matrix from a scRNA experiment where rows represent genes and
columns represent cells, and one wants to regress out mean
differences between genes.}

\item{row_size_factor}{Boolean indicating if a size factor should be
used to normalize the likelihood across rows. This may be
useful when \code{Y} is a matrix from a scRNA experiment where rows 
represent cells and columns represent genes.}

\item{col_intercept}{Boolean indicating if intercept term should be
fit for each column of \code{Y}. This may be useful when \code{Y}
is a matrix from a scRNA experiment where rows represent cells and
columns represent genes, and one wants to regress out mean
differences between genes.}
}
\value{
An object capturing the final state of the model fit. It will contain
the final values of \eqn{U}, \eqn{D} and \eqn{V}, as well as a list \code{progress}
that records information about the algorithm progress at each iteration.

An object capturing the initial state of the model fit. See
  \code{\link{fit_glmpca_pois}} for details.
}
\description{
Fit a Poisson GLM-PCA model to data matrix \code{Y}
  by maximum-likelihood estimation.
}
\details{
In generalized principal component analysis (GLM-PCA)
based on a Poisson likelihood (Townes et al, 2019), the counts
\eqn{y_{ij}} in the n x p matrix \eqn{Y} are modeled as
\deqn{y_{ij} \sim Poisson(\lambda_{ij}).} The logarithm of each
Poisson rate is defined as a linear combination of the parameters:
\deqn{\log \lambda_{ij} = \sum_{k=1}^K d_{k} u_{ik} v_{jk} + 
\sum_{m=1}^{n_{x}} x_{im} b_{jm} +
\sum_{l=1}^{n_{z}} w_{il} z_{jl} = (UDV' + XB' + WZ')_{ij},} where
\eqn{U} is an orthonormal \eqn{n \times K} matrix, \eqn{D} is a diagonal \eqn{K \times K} matrix,
\eqn{V} is an orthonormal \eqn{p \times K} matrix, \eqn{X} is a fixed \eqn{n \times n_{x}}
matrix of row specific covariates, \eqn{B} is an \eqn{p \times n_{x}} matrix
of coefficients of the row specific covariates, \eqn{Z} is a fixed 
\eqn{p \times n_{z}} matrix of column specific covariates, and
\eqn{W} is an \eqn{n \times n_{z}} matrix of coefficients of column
specific covariates. \eqn{K} is a tuning parameter specifying the rank of the matrices
\eqn{U} and \eqn{V}. \code{fit_glmpca_pois} computes maximum-likelihood
estimates (MLEs) of \eqn{U}, \eqn{D}, \eqn{V}, and if any row or column
covariates are present, respectively, \eqn{B} and \eqn{W}.

The algorithm works by repeatedly alternating between updating \eqn{U} and \eqn{W} with
\eqn{V} and \eqn{Z} fixed and updating
\eqn{V} and \eqn{Z} with \eqn{U} and \eqn{W} fixed. Each update takes the 
form of a series of Poisson regressions solved using cyclic co-ordinate
descent (ccd). When the algorithm
terminates, we rotate the columns of \eqn{U} and \eqn{V} 
so that they form an orthonormal set, and then we calculate \eqn{D} appropriately.
This rotation does not change the final log-likelihood.

The \code{control} argument is a list in which any of the following
named components will override the default optimization algorithm
settings (as they are defined by \code{fit_glmpca_control_default}):

\describe{

\item{\code{num_iter}}{Number of ccd updates to be made to parameters
  at each iteration of the algorithm.}

\item{\code{line_search}}{Boolean indicating if backtracking line search
  should be performed at each ccd iteration.}

\item{\code{alpha}}{alpha value of line search between 0 and .5.}

\item{\code{beta}}{beta value of line search between 0 and .5.}
  
\item{\code{calc_deriv}}{boolean indicated if maximum absolute derivatives of
 \eqn{U} and \eqn{V} should be calculated at each step of optimization. This may
 be useful for monitoring convergence though may have substantial computational
 cost for large matrices.}
  
\item{\code{calc_max_diff}}{boolean indicating if maximum absolute difference 
between successive updates of \eqn{U} and \eqn{V} should be calculated and
stored. This may be useful for monitoring convergence.}
  
}
}
\examples{
set.seed(1)

n <- 1000
p <- 500
K <- 1

data <- generate_glmpca_data_pois(n, p, K)
# initialize fit
fit0 <- init_glmpca_pois(
  Y = data$Y, 
  K = K,
  fit_col_size_factor = TRUE, 
  fit_row_intercept = TRUE
)

fit <- fit_glmpca_pois(Y = data$Y, fit0 = fit0)

}
\references{
Townes, F. W., Hicks, S. C., Aryee, M. J. and Irizarry,
R. A. (2019). Feature selection and dimension reduction for
single-cell RNA-Seq based on a multinomial model. \emph{Genome Biology}
\bold{20}, 295. \url{https://doi.org/10.1186/s13059-019-1861-6}

Collins, M., Dasgupta, S. and Schapire, R. E. (2002). A
generalization of principal components analysis to the exponential
family. In \emph{Advances in Neural Information Processing Systems} 14.
}
\seealso{
\code{\link{fit_glmpca_pois}}
}
