% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_glmpca_pois}
\alias{init_glmpca_pois}
\title{Initialize GLM-PCA Poisson Fit}
\usage{
init_glmpca_pois(
  Y,
  K,
  U,
  V,
  X,
  Z,
  B,
  W,
  fixed_b_cols,
  fixed_w_cols,
  fit_col_size_factor = TRUE,
  fit_row_intercept = TRUE,
  fit_row_size_factor = FALSE,
  fit_col_intercept = FALSE
)
}
\arguments{
\item{Y}{n x p matrix of counts; all entries should be
non-negative. Sparse matrices lead to faster computations.}

\item{K}{An integer 1 or greater giving the matrix rank. This
argument should only be specified if the initial fit (\code{U, V})
is not provided.}

\item{U}{An optional argument giving the initial estimate of the
loadings matrix. It should be an n x K matrix, where n is the
number of rows in the counts matrix \code{Y}, and K >= 1 is the rank
of the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{V}{An optional argument giving is the initial estimate of the
factors matrix. It should be a p x K matrix, where p is the number
of columns in the counts matrix \code{Y}, and K >= 1 is the rank of
the matrix factorization. When \code{U} and \code{V} are not
provided, input argument \code{K} should be specified instead.}

\item{X}{An optional argument giving row specific covariates of the 
count matrix \code{Y}. It should be n x n_x matrix, where n
is the number of rows of the count matrix \code{Y}, and
n_x is the number of row covariates.}

\item{Z}{An optional argument giving column specific covariates of the 
count matrix \code{Y}. It should be p x n_z matrix, where p
is the number of columns of the count matrix \code{Y}, and
n_z is the number of column covariates.}

\item{B}{An optional argument giving the initial estimates
for the coefficients of the row specific covariates of the 
count matrix \code{Y}. It should be p x n_x matrix, where p
is the number of columns of the count matrix \code{Y}, and
n_x is the number of row covariates.}

\item{W}{An optional argument giving the initial estimates
for the coefficients of the column specific covariates of the 
count matrix \code{Y}. It should be n x n_z matrix, where n
is the number of rows of the count matrix \code{Y}, and
n_z is the number of column covariates.}

\item{fixed_b_cols}{An optional argument giving which, if any, columns
of \code{B} should be fixed during optimization.}

\item{fixed_w_cols}{An optional argument giving which, if any, columns
of \code{W} should be fixed during optimization.}

\item{fit_col_size_factor}{Boolean indicating if a size factor should be
used to normalize the likelihood across columns. This may be
useful when \code{Y} is a matrix from a scRNA experiment where rows 
represent genes and columns represent cells.}

\item{fit_row_intercept}{Boolean indicating if intercept term should be fit
for each row of \code{Y}. This may be useful when \code{Y} is a matrix from a
scRNA experiment where rows represent genes and columns represent cells,
and one wants to regress out mean differences between genes.}

\item{fit_row_size_factor}{Boolean indicating if a size factor should be
used to normalize the likelihood across rows. This may be
useful when \code{Y} is a matrix from a scRNA experiment where rows 
represent cells and columns represent genes.}

\item{fit_col_intercept}{Boolean indicating if intercept term should be fit
for each column of \code{Y}. This may be useful when \code{Y} is a matrix from a
scRNA experiment where rows represent cells and columns represent genes,
and one wants to regress out mean differences between genes.}
}
\value{
An object capturing the initial state of the model fit. See
  \code{\link{fit_glmpca_pois}} for details.
}
\description{
Initialize a GLM-PCA Poisson model fit.
}
\seealso{
\code{\link{fit_glmpca_pois}}
}
