% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_glmpca_pois}
\alias{init_glmpca_pois}
\title{Initialize GLM-PCA Poisson Fit}
\usage{
init_glmpca_pois(
  Y,
  K,
  LL,
  FF,
  initialization = c("random", "expected_loglik"),
  fit_col_size_factor = FALSE,
  fit_row_intercept = FALSE,
  fixed_loadings = NULL,
  fixed_factors = NULL
)
}
\arguments{
\item{Y}{n x p matrix of counts; all entries should be
non-negative.}

\item{K}{An integer 1 or greater giving the matrix rank. This
argument should only be specified if the initial fit (\code{LL, FF})
is not provided.}

\item{LL}{An optional argument giving the initial estimate of the
loadings matrix. It should be an K x n matrix, where n is the
number of rows in the counts matrix \code{Y}, and K >= 1 is the rank
of the matrix factorization. When \code{LL} and \code{FF} are not
provided, input argument \code{K} should be specified instead.}

\item{FF}{An optional argument giving is the initial estimate of the
factors matrix. It should be a K x p matrix, where p is the number
of columns in the counts matrix \code{Y}, and K >= 1 is the rank of
the matrix factorization. When \code{LL} and \code{FF} are not
provided, input argument \code{K} should be specified instead.}

\item{fit_col_size_factor}{Boolean indicating if a size factor should be
used to normalize the likelihood across columns. This is done by fixing
an element of each loading to 1 and fixing and element of each factor
as the log the mean value of it's corresponding column. This may be
useful when \code{Y} is a matrix from a scRNA experiment where rows 
represent genes and columns represent cells.}

\item{fit_row_intercept}{Boolean indicating if intercept term should be fit
for each row of \code{Y}. This is done by fixing an element of each
factor to 1. This may be useful when \code{Y} is a matrix from a
scRNA experiment where rows represent genes and columns represent cells,
and one wants to regress out mean differences between genes.}

\item{fixed_loadings}{Vector of integers indicating, which, if any, loadings
should be fixed at their initial values. This argument will be ignored if
\code{LL} is not provided.}

\item{fixed_factors}{Vector of integers indicating which, if any, factors
should be fixed at their initial values. This argument will be ignored
if \code{FF} is not provided.}
}
\value{
An object capturing the initial state of the model fit. See
  \code{\link{fit_glmpca_pois}} for details.
}
\description{
Initialize a GLM-PCA Poisson model fit.
}
\seealso{
\code{\link{fit_glmpca_pois}}
}
