% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plash.R
\name{plash_omni}
\alias{plash_omni}
\title{Fit Poisson Matrix Factorization Algorithm}
\usage{
plash_omni(
  Y,
  K,
  offset = FALSE,
  intercept = FALSE,
  constant_offset = TRUE,
  parallel = TRUE,
  tol = 1e-08,
  update_L = T,
  update_F = T,
  update_c = T,
  update_L_c_simul = FALSE,
  init_LL = NULL,
  init_FF = NULL,
  init_cc = NULL,
  update_c_first = TRUE,
  min_iter = 3,
  max_iter = 30
)
}
\arguments{
\item{Y}{n x p data matrix}

\item{K}{number of factors to fit}

\item{offset}{boolean indicating if cell specific offset should be used}

\item{intercept}{boolean indicating if gene specific intercept should be used}

\item{constant_offset}{boolean indicating if c should be multiplied by cell specific offset 
in link function}

\item{parallel}{boolean indicating if the algorithm should be run in parallel}

\item{tol}{relative tolerance of log likelihood for convergence}

\item{update_L}{boolean indicating if L should be updated}

\item{update_F}{boolean indicating if F should be updated}

\item{update_c}{boolean indicating if c should be updated}

\item{update_L_c_simul}{boolean indicating if}

\item{init_LL}{}

\item{init_FF}{}

\item{init_cc}{}

\item{update_c_first}{}

\item{min_iter}{}

\item{max_iter}{}
}
\description{
Fit Poisson Matrix Factorization Algorithm
}
